+++ CHANGELOG.md
@@ -13,6 +13,7 @@
  - Fixed a panic when parsing a truncated language code in `iso_639_language_descriptor`
  - Fixed a panic when reported PES header length does not fit within available space
  - Fixed a panic due to a bug parsing PES header `ESCR_base` field
+ - Fixes a panic when `PES_header_data_length` is not long enough to accommodate all the headers actually present
 
 ## 0.14.0 - 2021-04-11
 ### Changed